Clean up Column Component View and Slot behaviours
