+++ src/Views/Traits/Helpers/ViewComponentColumnHelpers.php
@@ -4,22 +4,6 @@ namespace Rappasoft\LaravelLivewireTables\Views\Traits\Helpers;
 
 trait ViewComponentColumnHelpers
 {
-    /**
-     * Retrieves the defined Component View
-     */
-    public function getComponentView(): string
-    {
-        return $this->componentView;
-    }
-
-    /**
-     * Determines whether a Component View has been set
-     */
-    public function hasComponentView(): bool
-    {
-        return isset($this->componentView);
-    }
-
     public function hasCustomComponent(): bool
     {
         return isset($this->customComponentView);
